feat: fixing build errors, adding metadata API handler
