+++ embed/package.json
@@ -20,6 +20,10 @@
     "lint:fix": "eslint --fix --ext .ts,.js,.tsx ."
   },
   "dependencies": {
+    "@gitcoin/passport-identity": "^1.0.0",
+    "@gitcoin/passport-platforms": "^1.0.0",
+    "@gitcoin/passport-types": "^1.0.0",
+    "@spruceid/didkit-wasm-node": "^0.2.1",
     "axios": "^0.27.2",
     "cors": "^2.8.5",
     "dotenv": "^16.0.0",
@@ -27,11 +31,7 @@
     "express": "4",
     "luxon": "^2.4.0",
     "tslint": "^6.1.3",
-    "uuid": "^8.3.2",
-    "@gitcoin/passport-identity": "^1.0.0",
-    "@gitcoin/passport-platforms": "^1.0.0",
-    "@gitcoin/passport-types": "^1.0.0",
-    "@spruceid/didkit-wasm-node": "^0.2.1"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.23.3",
@@ -46,16 +46,19 @@
     "@types/node-fetch": "latest",
     "@types/supertest": "^2.0.12",
     "@types/webpack-env": "^1.16.3",
+    "@typescript-eslint/parser": "^6.19.0",
     "babel-jest": "^29.7.0",
     "babel-plugin-replace-import-extension": "^1.1.3",
     "babel-plugin-transform-import-meta": "^2.2.1",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint": "^8.52.0",
     "jest": "^29.6.4",
+    "prettier": "^3.1.0",
+    "prettier-eslint": "^16.1.2",
     "supertest": "^6.2.2",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tsx": "^4.7.0",
-    "typescript": "^5.3.3"
+    "typescript": "^4.4.4"
   },
   "resolutions": {
     "leveldown": "6.1.1"
