+++ iam/package.json
@@ -31,6 +31,7 @@
     "@gitcoin/passport-identity": "^1.0.0",
     "@gitcoin/passport-platforms": "^1.0.0",
     "@gitcoin/passport-types": "^1.0.0",
+    "@ipld/dag-cbor": "^7.0.3",
     "@spruceid/didkit-wasm-node": "^0.2.1",
     "multiformats": "^13.0.1",
     "axios": "^0.27.2",
@@ -41,10 +42,8 @@
     "ethers": "^6.13.4",
     "express": "4",
     "google-auth-library": "^7.14.1",
-    "@ipld/dag-cbor": "^7.0.3",
     "luxon": "^2.4.0",
     "moralis": "^2.24.2",
-    "multiformats": "^9.9.0",
     "tslint": "^6.1.3",
     "twitter-api-v2": "^1.15.1",
     "uuid": "^8.3.2"
@@ -62,11 +61,14 @@
     "@types/node-fetch": "latest",
     "@types/supertest": "^2.0.12",
     "@types/webpack-env": "^1.16.3",
+    "@typescript-eslint/parser": "^6.19.0",
     "babel-jest": "^29.7.0",
     "babel-plugin-replace-import-extension": "^1.1.3",
     "babel-plugin-transform-import-meta": "^2.2.1",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint": "^8.52.0",
     "jest": "^29.6.4",
+    "prettier": "^3.1.0",
+    "prettier-eslint": "^16.1.2",
     "supertest": "^6.2.2",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
