+++ platforms/src/index.ts
@@ -2,6 +2,7 @@
 import platforms from "./platforms";
 import { createProviders } from "./utils/createProviders";
 import { keccak256, toUtf8Bytes } from "ethers";
+import { PlatformGroupSpec } from "./types";
 
 // Check that all platforms have a ProviderConfig, PlatformDetails, and providers
 Object.entries(platforms).map(([platformName, platform]) => {
@@ -22,6 +23,69 @@ Object.values(platforms).map(({ ProviderConfig }) => {
 
 // This is used in tests & IAM only, not in the app
 export const providers = createProviders(platforms);
+
+const skipPlatforms = ["ClearText"];
+
+type StampData = {
+  name: string;
+  description: string;
+  hash: string;
+};
+
+type GroupData = {
+  name: string;
+  stamps: StampData[];
+};
+
+type PlatformData = {
+  name: string;
+  icon: string;
+  description: string;
+  connectMessage: string;
+  groups: GroupData[];
+};
+
+const formatPlatformGroups = (providerConfig: PlatformGroupSpec[]) =>
+  providerConfig.reduce(
+    (groups: GroupData[], group: PlatformGroupSpec) => [
+      ...groups,
+      {
+        name: group.platformGroup,
+        stamps: group.providers.map(({ name, title, hash }) => {
+          if (!hash) {
+            throw new Error(`No hash defined for ${name}`);
+          }
+          return {
+            name,
+            hash,
+            description: title,
+          };
+        }),
+      },
+    ],
+    [] as GroupData[]
+  );
+
+export const platformsData = Object.entries(platforms).reduce((data, [id, platform]) => {
+  if (skipPlatforms.includes(id)) return data;
+
+  const { name, icon, description, connectMessage } = platform.PlatformDetails;
+  if (!icon) throw new Error(`No icon defined for ${id}`);
+
+  const groups = formatPlatformGroups(platform.ProviderConfig);
+
+  return [
+    ...data,
+    {
+      id,
+      name,
+      icon,
+      description,
+      connectMessage,
+      groups,
+    },
+  ];
+}, [] as PlatformData[]);
 
 export * from "./types";
 export { Platform as PlatformClass } from "./utils/platform";
